feat: Update for Floresta example

+++ src/types/mod.rs
@@ -4,6 +4,6 @@ pub use responses::*;
 #[derive(Debug, serde::Serialize)]
 pub enum HashOrHeight {
     Hash(bitcoin::BlockHash),
-    Height(u32),
+    Height(i64),
 }
 pub type FeeRate = bitcoin_units::FeeRate;
